+++ demo/src/main/java/com/example/demo/repositorio/CarritoComprasRepository.java
@@ -6,6 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CarritoComprasRepository extends JpaRepository<CarritoCompras, Integer> {
-
-    CarritoCompras findByClienteModelId(Integer clienteId);   // <---- AGREGA ESTE MÉTODO
+    CarritoCompras findByClienteModelId(Long clienteId);
 }
